writePass/printPass w/o YAXA
Beginning of a conversion of the program from using the YAXA stream cipher in the first cascade algorithm, to using OpenSSL's EVP interface to perform this encryption.  Currently writePass() and printPasses() work using blowfish in OFB mode, though writePass() should NOT be working with EVP_BytestoKey (it's mismatching between evpCipher1 and evpCipher2, evpKey1 and evpKey2, etc).  Meanwhile updateEntry() is using the same construct in printPasses() to decode the file into decryptedBuffer, but the first EVP_DeryptUpdate() function in updateEntry() causes a segfault.  valgrind claims this says something about "Address 0x12 is not stack'd, malloc'd or (recently) free'd"