1st Algorithm in Cascade Selectable
The 1st algorithm in the cascade can now be selected as the 2nd could.  The message digest used for the key derivation can also be specified.  Error checking is also now done to ensure that only OFB, CFB or CTR modes are selected.  

AEAD of the 2nd cipher-text was removed and limited to AE of the 1st algorithm's plain-text. (MAC-and-encrypt)