Update to 2.1.0
Authentication Composition changed to Encrypt-then-MAC from MAC-and-encrypt.  Previously MAC-and-encrypt was used, because during testing I found that dumping malicious data into the temporary files lead to corruption of the password database.  Upon further thinking, if there were an attacker live on your system that could inject data into the temporary files with an intent to damage the password database, then they could probably just delete the database.  Since other types of data corruption are always possible, it shouldn't really be up to this program to protect the from data corruption, and it should just "stay in its lane" and prioritize secrecy.  With that in mind, Encrypt-then-MAC is the best practice by a wide consensus, with only Schneier/Ferguson recommending MAC-then-encrypt instead.

Quick help summary was changed to reflect the ability to choose both algorithms in the cascade.  The ciphers and digests will now be selected by parsing a string of "first-cipher:second-cipher" or "first-digest:second:digest" given to -c or -H respectively.