Changed the way sendwithxclip.c exits after one paste
The previous method used to test if a selection had been pasted was not accurate, so adopted another method that exits 1 ms after the password was pasted, effectively meaning the password is immediately pasted without the user able to paste it more than once. One benefit of this behavior is that it can later be modified to allow the user to paste it any of number of times, so long as the interim betwee pastings is shorter than the timeout limit.