Using XPending method to exit sendWithXlib instead of a timer thread
The method which tests XPending was not working because the timer was too short at only 1 ms.  Some OS's and windows and combinations of each require a different amount of time for the password to paste and for XPending to test false accurately. Therefore instead of exiting 1ms after XPending tests false, the default 30 seconds is used to give it enough time to paste the password into any window regardless of how many X calls it needs.  This also gives the user the opportunity to paste the password more than once before it is cleared, and also allows them to not rush to paste the password before the specified amount of seconds because the timer will not start until a paste is made.