Updated documentation and information prompts
Program usage and manual files were updated to reflect ability to send to different X selections, and informational messages printed by the program were changed to reflect behavior of when the password will be cleared.  For some reason, the timer that begins only after the password is pasted only works if using the 'primary' selection, so if using the 'clipboard' selection, the timer will start immediately as if piping out to xclip binary.