Adding password verification
A third MAC generated by HMAC on the user-supplied password is used to test if the password entered was correct.  This way the program can differentiate between the database not passing authentication because it was modified, or if the password was not correct, and inform the user.  Before either case would trip the same authentication failure message and leave the user guessing.