Adding password checking
Adding a method to differentiate between an incorrect password and a MAC failure. This method encrypts the salt to the end of the file, and so then compares a decryption of the salt to check if the password decrypts successfully. Currently it's only using aes-256-ctr to do this part instead of the user chosen cipher for the rest of the ciphertext because it will require the handling of unpredictable file lengths with block modes. I'm also a little skeptical that adding a cipher-text that has a known plain-text to the end of the file is a good idea.