Password checking, with user-configured cipher
Added ability to deal with unpredictable file-length depending on what cipher the user chooses.  This way the salt can be encrypted and appended to the end of the file, and a trial decryption can be made to check if the password used was correct. If it was, the decryption will match the salt, if not, it won't.  Previously this encryption could only be done with aes-256-ctr, without respect to which cipher the user chose, because using a block cipher would make determining how far to the end of the file unpredictable.  Now it can use EVP_MAX_BLOCK_LENGTH and EVP_get_cipher_block_length() to determine this and respect the user's cipher choice. However, it will cause EVP_Decrypt_FinalEx() to return unsuccessfully, so a new function using this had to be used so that the program doesn't return an error if it is unsuccessful, but can still compare the result of the decryption against the salt to verify the password.

Also verified that adding this cipher-text of the known plain-text salt isn't a major concern because it can only reveal information about the one-time keystream (a product of the key, iv and nonces) instead of the key itself.  Also becuse the salt is only 32-bytes, it doesn't provide enough information to be useful.

Line 2292 was also erroneously changed to always use aes-256-ctr and has been changed back.