Updates to authenticity/integrity checks and other changes
The database checksum was added to the end of the database so that it can hash the keyed hash for the database password as well as the cipher-text/IV MAC.
New scripts were added to simulate a corrupt or forged database.
Documentation was updated.
General changes to variable names and code comments were changed.