Changed way some associated data is checked and error handling
Instead of using NID numbers to verify cipher and message digests specificed as associated data, the actual algorithm names themselves are now used.

Also changed the way errorr handling was done, with some goto statements so that the same block of free statements doesn't have to be made over and over.

Fixed some memory leaks and unsantized buffers as well