Scrypt version
As of version 1.1.1 OpenSSL has scrypt and as of 1.1.0 it has HKDF. This version derives a master key with scrypt, and then uses HKDF to derive an encryption key, and authentication key and an IV for encryption. It also does away with PBKDF2 and thus the ability/requirement of the user to specify a message digest, but the code for this hasn't been removed at this point because I haven't decided if I want to make it optional to use PBKDF2 or scrypt.  The next step will be to make the scrypt work factors configurable.