Update to use scrypt and new key derivation scheme
The program has been updated to use scrypt instead of PBKDF2, and also changed in how it derives keys for the cipher algorithm and authentication. Scrypt is now used to derive a 1024 bit master key that is then split in half.  One half is supplied to HMAC as the authentication key, and the other half is used for the specified cipher algorithm, with any surplus bits being discharded. The IV for encryption is also now replaced by the salt since it serves the same purpose.  The IV could have been ommitted but since the OpenSSL functions epect it and it was trivial to just use the salt in place of the IV that is what was done.  However, the authentication of association data now only considers the salt instead of the salt and IV.  Other improvements were made to the option handling to make it more clear when incorrect syntax was passed, when a file is needed, etc.